requirements.txt
faster-whisper==1.0.3
openai>=1.0.0
pyaudio
python-dotenv
